[android] Added Savalbe interface and moved Detachable interface to base package
