+++ android/src/com/mapswithme/maps/base/Savable.java
@@ -1,0 +1,9 @@
+package com.mapswithme.maps.base;
+
+import android.support.annotation.NonNull;
+
+public interface Savable<T>
+{
+  void onSave(@NonNull T outState);
+  void onRestore(@NonNull T inState);
+}
